Create Final Discounted Price.cpp
A shopkeeper has a sale to complete and has arranged the items being sold in an array. Starting from the left, the shopkeeper rings up each item at its full price less the price of the first lower or equal priced item to its right. If there is no item to the right that costs less than or equal to the current item's price, the current item is sold at full price. For example, assume there are items priced [2,3,1,2,4,2]
.

The items 0
 and 1
 are each discounted by 1
 unit, the first equal or lower price to the right.
Item 2
, priced 1
 unit sells at full price because there are no equal or lower priced items to the right.
The next item, item 3
 at 2
 units, is discounted 2
 units to 0
.
Item 4
 at 4
 units is discounted 2
 units as well, so its final price is 2
.
The final item 5
 at 2
 units must be purchased at full price because there are no lower prices to the right.
The total cost is 1+2+1+0+2+2=8
 units. The full price items are at indices [2,5]
 using 0
 based indexing.
Input
The first line contains n
, the number of items.
The next n
 lines contain the elements of prices
, the array of item prices.
Constraints
1≤n≤105
1≤prices[i]≤106
, where 0≤i<n
Output
Print the total cost of all the items on the first line.
On the second line, print a space-separated list of integers representing the indices of the items purchased at full price, in ascending index order.
Examples
inputCopy
6
5
1
3
4
6
2
outputCopy
14
1 5 
inputCopy
5
1
3
3
2
5
outputCopy
9
0 3 4 
Note
Explanation for test case 1 The prices of the items are given by the array:

index     0 1 2 3 4 
prices = [1,3,3,2,5]
We can find the discount on each item:
prices[0] = 1, no lower or equal price to the right, no discount price = 1
prices[1] = 3, first lower or equal to the right = prices[2] = 3, discounted price = 3-3=0
prices[2] = 3, first lower or equal to the right = prices[3] = 2, discounted price = 3-2 = 1
prices[3] = 2, no lower or equal price to the right, no discount price = 2
prices[4] = 5, no lower or equal price to the right, no discount, price = 5
The final discounted price is 1+0+1+2+5= 9. There is no discount for item[0], item[3] or item[4].
Explanation for test case 2 The prices of the items are given by the array:

index     0 1 2 3 4 5
prices = [5,1,3,4,6,2]
We can find the discount on each item:
prices[0] = 5. first lower or equal price to the right = prices[1]= 1. discounted price =5-1=4
prices[1] = 1, lowest to the right = prices[5] = 2, no lower price to the right so no discount, price = 1
prices[2] = 3, first lower or equal = prices[5] = 2, discounted price = 3-2 = 1
prices[3] = 4, first lower or equal = prices[5] = 2, discounted price = 4-2 = 2
prices[4] = 6, first lower or equal = prices[5] = 2. discounted price = 6-2 = 4
prices[5] = 2, no lower or equal to the right, so no discount price = 2
The final discounted price is 4+1+1+2+4+2 = 14. There is no discount for item[1] or item[5].